Add configuration
This is a first pass at what the configuration will look like. I expect it will
change as we actually implement the program, of course.

There is a sample configuration file and a test that it deserializes into the
expected Go structures.

This isn't hooked up to anything yet, except logging the config and exiting.