Run a webserver
This runs a simple webserver to host the demonstration certificates.

Right now, there's a placeholder which loads a key pair from files.
This is going to be deleted in an upcoming PR, but is enough to get
started without making this PR too big.

The page content is also a plain text message, pending adding a better
page content to be rendered.