Add certificate storage
This is code to store keys and certificates on disk.
The general way I expect this to be used is outlined in the test:
A user generates a key, goes off to a CA for a cert, and stores it as the
"next" certificate. Once it is verified to be "good" for whatever definition
that means to the caller, they can copy it to the current key.

The goal here is to abstract the disk format away from the user, but it uses
PEM for both PKCS8 keys and certificates to make it easy to use the files in
an external program if required (eg, openssl x509 -text ...)