Load Certificates from storage (#11)
This loads certificates from storage, instead of the temporary hack that
was in place.

It moves test configs into a testdata folder, per go norms

It adds a public LoadCertificate function, which is currently used at
startup to load certs, and will be used to reload when the ACME client
has completed issuance.