Set up an ACME client (#14)
This sets up an ACME client.

There's a nontrivial amount of boilerplate here required for using Lego,
like the legoUser struct.

The Run() function doesn't do anything yet, but that's where the actual
issuance loop logic will go.

The storage package needed to be extended to store the ACME account.
While I could have just saved the file in the ACME client, I think it
makes sense to keep all the filesystem interaction in one place,
especially if or when we want to make this able to use some type of
external secrets storage systems.

I removed the CACerts config because Lego supports using
LEGO_CA_CERTIFICATES out of the box, so I can point that at a
pebble.minica.pem.