Add a scheduler (#16)
The basic API here is to make a new scheduler with scheduler.New()
It calls loop() in a goroutine.

RunAt and RunIn are the ways to add new jobs.

Stop can be run to stop processing jobs (eg, on shutdown).

test-certs-site will need to do tasks on a schedule:

* Check ARI to see if a renewal is needed
* Wait for a certificate to expire
* Re-check CRLs waiting for revocation

I had started doing those tasks with ad-hoc goroutines and sleeps. But I
realized that having a schedule as a primitive would make it much easier
to defer tasks into the future, and (not present here) to have consistent
monitoring of when those tasks run.

The most annoying part of this is using golang's container/heap, which
has a bit of boilerplate. That code should all go away if/when Golang adds a
generic heap to containers.